+++ packages/front/src/hooks/mutation/expense.ts
@@ -1,13 +1,22 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query'
-
 import { addExpenseApi, deleteExpenseApi, updateExpenseApi } from '@/client/expenses'
-import { useSetRecoilState } from 'recoil'
-import { expenseAddIsFetchingState } from '@/store/expenseFetchingState'
+import { useRecoilValue, useSetRecoilState } from 'recoil'
+import { expenseAddIsFetchingState, expenseDeleteQueue, expenseEditQueue } from '@/store/expenseFetchingState'
 import { EXPENSES } from '@/constants/query'
+import { searchState, sortState } from '@/store/filterState'
+import { ExpensesListData } from '../useExpensesListQuery'
 
+/**
+ * @note SSRSuspense(components/common에 있는 컴포넌트)로 감싸지 않은 컴포넌트에서 useQuery를 사용할 수 없기 때문에 mutation을 분리합니다.
+ */
 export const useExpensesListMutation = () => {
   const queryClient = useQueryClient()
+  const setExpenseEditQueue = useSetRecoilState(expenseEditQueue)
+  const setExpenseDeleteQueue = useSetRecoilState(expenseDeleteQueue)
   const setIsFetching = useSetRecoilState(expenseAddIsFetchingState)
+
+  const searchQuery = useRecoilValue(searchState)
+  const sortQuery = useRecoilValue(sortState)
 
   const errorHandler = (res: Response) => {
     if (res.status === 401) alert('접근 권한이 없습니다.')
@@ -20,7 +29,8 @@ export const useExpensesListMutation = () => {
     onError: errorHandler,
     onSuccess: () => {
       queryClient.invalidateQueries({
-        queryKey: [EXPENSES]
+        queryKey: [EXPENSES],
+        exact: false
       })
     },
     onSettled: () => {
@@ -30,28 +40,42 @@ export const useExpensesListMutation = () => {
 
   const updateExpense = useMutation({
     mutationFn: updateExpenseApi,
-    onSuccess: () => {
+    onSuccess: (_data, variable) => {
       queryClient.invalidateQueries({
-        queryKey: [EXPENSES]
+        queryKey: [EXPENSES],
+        exact: false
       })
+      setExpenseEditQueue((prev) => prev.filter((id) => id !== variable.id))
     },
-    onError: errorHandler,
-    onSettled: () => {
-      setIsFetching(false)
-    }
+    onError: errorHandler
   })
 
   const deleteExpense = useMutation({
     mutationFn: deleteExpenseApi,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: [EXPENSES]
-      })
+    onMutate: (expenseId) => {
+      queryClient.setQueryData(
+        [EXPENSES, searchQuery, sortQuery],
+        (prev: { pages: ExpensesListData[]; pageParams: number[] }) => {
+          console.log('PREV DATA :', prev)
+          const pages = prev.pages.map((page) => ({
+            ...page,
+            content: page.content.filter((expense) => expense.id !== expenseId)
+          }))
+          return {
+            ...prev,
+            pages
+          }
+        }
+      )
     },
-    onError: errorHandler,
-    onSettled: () => {
-      setIsFetching(false)
-    }
+    onSuccess: (_data, expenseId) => {
+      queryClient.invalidateQueries({
+        queryKey: [EXPENSES],
+        exact: false
+      })
+      setExpenseDeleteQueue((prev) => prev.filter((id) => id !== expenseId))
+    },
+    onError: errorHandler
   })
 
   return {
