+++ packages/front/src/hooks/useExpensesListQuery.ts
@@ -1,11 +1,9 @@
 import { fetcher } from '@/client/fetcher'
 import { Expenses } from '@/model'
-import { useMutation, useSuspenseInfiniteQuery, useQueryClient } from '@tanstack/react-query'
+import { useSuspenseInfiniteQuery } from '@tanstack/react-query'
 import { useEffect, useRef } from 'react'
-import { addExpenseApi, deleteExpenseApi, updateExpenseApi } from '@/client/expenses'
 import { useIntersectionObserver } from 'greenbean-pack'
-import { useRecoilValue, useSetRecoilState } from 'recoil'
-import { expenseDeleteQueue, expenseEditQueue } from '@/store/expenseFetchingState'
+import { useRecoilValue } from 'recoil'
 import { EXPENSES } from '@/constants/query'
 import { searchState, sortState } from '@/store/filterState'
 
@@ -18,9 +16,6 @@ export interface ExpensesListData {
 }
 
 export const useExpensesListInfiniteQuery = () => {
-  const queryClient = useQueryClient()
-  const setExpenseEditQueue = useSetRecoilState(expenseEditQueue)
-  const setExpenseDeleteQueue = useSetRecoilState(expenseDeleteQueue)
   const searchQuery = useRecoilValue(searchState)
   const sortQuery = useRecoilValue(sortState)
   const { data, fetchNextPage, isLoading, isFetching, isFetchingNextPage } = useSuspenseInfiniteQuery<ExpensesListData>(
@@ -45,59 +40,6 @@ export const useExpensesListInfiniteQuery = () => {
     if (entry) fetchNextPage()
   }, [entry])
 
-  const errorHandler = (error: Error) => {
-    process.env.NODE_ENV === 'development' && console.error(error)
-    alert(error.message)
-  }
-
-  const addExpense = useMutation({
-    mutationFn: addExpenseApi,
-    onError: errorHandler,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: [EXPENSES]
-      })
-    }
-  })
-
-  const updateExpense = useMutation({
-    mutationFn: updateExpenseApi,
-    onSuccess: (_data, variable) => {
-      queryClient.invalidateQueries({
-        queryKey: [EXPENSES]
-      })
-      setExpenseEditQueue((prev) => prev.filter((id) => id !== variable.id))
-    },
-    onError: errorHandler
-  })
-
-  const deleteExpense = useMutation({
-    mutationFn: deleteExpenseApi,
-    onMutate: (expenseId) => {
-      queryClient.setQueryData(
-        [EXPENSES, searchQuery, sortQuery],
-        (prev: { pages: ExpensesListData[]; pageParams: number[] }) => {
-          console.log('PREV DATA :', prev)
-          const pages = prev.pages.map((page) => ({
-            ...page,
-            content: page.content.filter((expense) => expense.id !== expenseId)
-          }))
-          return {
-            ...prev,
-            pages
-          }
-        }
-      )
-    },
-    onSuccess: (_data, expenseId) => {
-      queryClient.invalidateQueries({
-        queryKey: [EXPENSES]
-      })
-      setExpenseDeleteQueue((prev) => prev.filter((id) => id !== expenseId))
-    },
-    onError: errorHandler
-  })
-
   const expenseList = data.pages.flatMap((page) => [...page.content])
 
   return {
@@ -105,9 +47,6 @@ export const useExpensesListInfiniteQuery = () => {
     triggerRef: ref,
     isFetching,
     isFetchingNextPage,
-    isLoading,
-    updateExpenseMutate: updateExpense.mutate,
-    deleteExpenseMutate: deleteExpense.mutate,
-    addExpenseMutate: addExpense.mutate
+    isLoading
   }
 }
